+++ src/index.ts
@@ -9,6 +9,7 @@ export * from "./functions/clamp-number.js";
 export * from "./functions/escape-sql-like-string.js";
 export * from "./functions/format-cents-as-usd.js";
 export * from "./functions/format-dollars-as-usd.js";
+export * from "./functions/generate-random-integer.js";
 export * from "./functions/generate-random-string.js";
 export * from "./functions/heartbeat.js";
 export * from "./functions/is-local-ip-address.js";